API/requirements.txt
blinker==1.9.0
click==8.2.2
colorama==0.4.6
Flask==3.1.1
Flask-MySQLdb==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
Werkzeug==3.1.3
